requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coloredlogs==15.0.1
cryptography==42.0.5
Django==5.0.3
django-admin==2.0.2
django-enumfields==2.1.1
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
humanfriendly==10.0
idna==3.6
isoweek==1.3.3
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
razorpay==1.4.1
requests==2.31.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tzlocal==5.2
urllib3==2.2.1
xlwt==1.3.0
